Move contactsyncers to opengever.api.kub.
preparing a reuse in the dossier api deserialization.
